bsc_nat_fsm: tweak get_peer_addr
Rename the function from sccp_sap_get_peer_addr_in to get_peer_addr to
make the code using it slightly more readable.  The _in at the end of
the function is not needed anymore, this was used to differentiate from
peer_addr_out which was used in the caller code before the connection
mapping was implemented in I1556aa665fbb0a97507f98794e74820731fa6935.
Now instead of peer_addr_out, subscr_conn->bsc->addr or
subscr_conn->msc->addr are used.

Rename variable arguments to sccp_inst and addr to be consistent with
the variable names in the callers.

Related: SYS#5560
Change-Id: Ie023360724254be54cbaac4490b0341dfe68399f